- move product route to module
