+++ resources/application.conf
@@ -4,6 +4,8 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [ com.aditmodhvadia.ApplicationKt.main, com.aditmodhvadia.modules.users.UserModuleKt.user ]
+        modules = [ com.aditmodhvadia.ApplicationKt.main,
+         com.aditmodhvadia.modules.users.UserModuleKt.user ,
+         com.aditmodhvadia.modules.users.ProductModuleKt.product]
     }
 }
