+++ src/modules/products/service/InMemoryProductService.kt
@@ -1,7 +1,7 @@
-package com.aditmodhvadia.routes.products.service
+package com.aditmodhvadia.modules.products.service
 
 import com.aditmodhvadia.models.Product
-import com.aditmodhvadia.routes.products.data.ProductDataSource
+import com.aditmodhvadia.modules.products.data.ProductDataSource
 
 class InMemoryProductService(private val productDataSource: ProductDataSource) : ProductService {
     override fun findAll(): Collection<Product> = productDataSource.retrieveProducts()
