+++ test/modules/products/ProductRouteTest.kt
@@ -1,7 +1,9 @@
 package com.aditmodhvadia.routes.products
 
+import com.aditmodhvadia.installGson
 import com.aditmodhvadia.models.Product
 import com.aditmodhvadia.main
+import com.aditmodhvadia.modules.products.product
 import com.google.gson.GsonBuilder
 import io.ktor.http.*
 import io.ktor.server.testing.*
@@ -21,7 +23,7 @@ internal class ProductRouteTest {
         @Test
         fun `should return collection of products with success response`() {
             // when, then
-            withTestApplication({ main(testing = true) }) {
+            withTestApplication({ product(testing = true).apply { installGson() } }) {
                 handleRequest(HttpMethod.Get, "/products").apply {
                     assertThat(response.status()).isEqualTo(HttpStatusCode.OK)
                     assertThat(response.contentType().toString())
@@ -50,7 +52,7 @@ internal class ProductRouteTest {
             val productId = 1
 
             // when/then
-            withTestApplication({ main(testing = true) }) {
+            withTestApplication({ product(testing = true).apply { installGson() } }) {
                 handleRequest(HttpMethod.Get, "/products/$productId").apply {
                     assertThat(response.status()).isEqualTo(HttpStatusCode.OK)
                     assertThat(response.contentType().toString())
@@ -72,7 +74,7 @@ internal class ProductRouteTest {
             val productId = -1
 
             // when/then
-            withTestApplication({ main(testing = true) }) {
+            withTestApplication({ product(testing = true).apply { installGson() } }) {
                 handleRequest(HttpMethod.Get, "/products/$productId").apply {
                     assertThat(response.status()).isEqualTo(HttpStatusCode.NoContent)
                     assertThat(response.content).isEqualTo("Product with given id not found")
@@ -86,7 +88,7 @@ internal class ProductRouteTest {
             val productId = "invalid_product_id"
 
             // when/then
-            withTestApplication({ main(testing = true) }) {
+            withTestApplication({ product(testing = true).apply { installGson() } }) {
                 handleRequest(HttpMethod.Get, "/products/$productId").apply {
                     assertThat(response.status()).isEqualTo(HttpStatusCode.BadRequest)
                     assertThat(response.content).isEqualTo("Specified product id is incorrect")
